Move DictionaryEntry to shared
