+++ Bridge/Bridge.csproj
@@ -69,6 +69,7 @@
     <Compile Include="shared\System\AttributeTargets.cs" />
     <Compile Include="shared\System\AttributeUsageAttribute.cs" />
     <Compile Include="shared\System\CharEnumerator.cs" />
+    <Compile Include="shared\System\Collections\DictionaryEntry.cs" />
     <Compile Include="shared\System\Collections\Generic\HashSet.cs" />
     <Compile Include="shared\System\Collections\Generic\IComparer.cs" />
     <Compile Include="shared\System\Collections\Generic\IEnumerator.cs" />
@@ -220,7 +221,6 @@
     <Compile Include="System\Collections\Generic\IReadOnlyCollection.cs" />
     <Compile Include="shared\System\Collections\Generic\Queue.cs" />
     <Compile Include="shared\System\Collections\Generic\Stack.cs" />
-    <Compile Include="System\Collections\DictionaryEntry.cs" />
     <Compile Include="System\Collections\HashHelpers.cs" />
     <Compile Include="System\Collections\IDictionary.cs" />
     <Compile Include="System\ComponentModel\BrowsableAttribute.cs" />
