+++ Bridge/Resources/.generated/bridge.meta.js
@@ -189,6 +189,7 @@
     $m("System.Globalization.NumberFormatInfo", function () { return {"att":1048833,"a":2,"m":[{"a":2,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":"Clone","t":8,"sn":"clone","rt":$n[0].Object},{"a":2,"n":"GetFormat","t":8,"pi":[{"n":"formatType","pt":Function,"ps":0}],"sn":"getFormat","rt":$n[0].Object,"p":[Function]},{"a":2,"n":"CurrencyDecimalDigits","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_CurrencyDecimalDigits","t":8,"rt":$n[0].Int32,"fg":"currencyDecimalDigits","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_CurrencyDecimalDigits","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"currencyDecimalDigits"},"fn":"currencyDecimalDigits"},{"a":2,"n":"CurrencyDecimalSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_CurrencyDecimalSeparator","t":8,"rt":$n[0].String,"fg":"currencyDecimalSeparator"},"s":{"a":2,"n":"set_CurrencyDecimalSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"currencyDecimalSeparator"},"fn":"currencyDecimalSeparator"},{"a":2,"n":"CurrencyGroupSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_CurrencyGroupSeparator","t":8,"rt":$n[0].String,"fg":"currencyGroupSeparator"},"s":{"a":2,"n":"set_CurrencyGroupSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"currencyGroupSeparator"},"fn":"currencyGroupSeparator"},{"a":2,"n":"CurrencyGroupSizes","t":16,"rt":$n[0].Array.type(System.Int32),"g":{"a":2,"n":"get_CurrencyGroupSizes","t":8,"rt":$n[0].Array.type(System.Int32),"fg":"currencyGroupSizes"},"s":{"a":2,"n":"set_CurrencyGroupSizes","t":8,"p":[$n[0].Array.type(System.Int32)],"rt":$n[0].Void,"fs":"currencyGroupSizes"},"fn":"currencyGroupSizes"},{"a":2,"n":"CurrencyNegativePattern","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_CurrencyNegativePattern","t":8,"rt":$n[0].Int32,"fg":"currencyNegativePattern","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_CurrencyNegativePattern","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"currencyNegativePattern"},"fn":"currencyNegativePattern"},{"a":2,"n":"CurrencyPositivePattern","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_CurrencyPositivePattern","t":8,"rt":$n[0].Int32,"fg":"currencyPositivePattern","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_CurrencyPositivePattern","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"currencyPositivePattern"},"fn":"currencyPositivePattern"},{"a":2,"n":"CurrencySymbol","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_CurrencySymbol","t":8,"rt":$n[0].String,"fg":"currencySymbol"},"s":{"a":2,"n":"set_CurrencySymbol","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"currencySymbol"},"fn":"currencySymbol"},{"a":2,"n":"CurrentInfo","is":true,"t":16,"rt":$n[1].NumberFormatInfo,"g":{"a":2,"n":"get_CurrentInfo","t":8,"rt":$n[1].NumberFormatInfo,"fg":"currentInfo","is":true},"fn":"currentInfo"},{"a":2,"n":"InvariantInfo","is":true,"t":16,"rt":$n[1].NumberFormatInfo,"g":{"a":2,"n":"get_InvariantInfo","t":8,"rt":$n[1].NumberFormatInfo,"fg":"invariantInfo","is":true},"fn":"invariantInfo"},{"a":2,"n":"NaNSymbol","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_NaNSymbol","t":8,"rt":$n[0].String,"fg":"nanSymbol"},"s":{"a":2,"n":"set_NaNSymbol","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"nanSymbol"},"fn":"nanSymbol"},{"a":2,"n":"NegativeInfinitySymbol","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_NegativeInfinitySymbol","t":8,"rt":$n[0].String,"fg":"negativeInfinitySymbol"},"s":{"a":2,"n":"set_NegativeInfinitySymbol","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"negativeInfinitySymbol"},"fn":"negativeInfinitySymbol"},{"a":2,"n":"NegativeSign","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_NegativeSign","t":8,"rt":$n[0].String,"fg":"negativeSign"},"s":{"a":2,"n":"set_NegativeSign","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"negativeSign"},"fn":"negativeSign"},{"a":2,"n":"NumberDecimalDigits","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_NumberDecimalDigits","t":8,"rt":$n[0].Int32,"fg":"numberDecimalDigits","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_NumberDecimalDigits","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"numberDecimalDigits"},"fn":"numberDecimalDigits"},{"a":2,"n":"NumberDecimalSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_NumberDecimalSeparator","t":8,"rt":$n[0].String,"fg":"numberDecimalSeparator"},"s":{"a":2,"n":"set_NumberDecimalSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"numberDecimalSeparator"},"fn":"numberDecimalSeparator"},{"a":2,"n":"NumberGroupSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_NumberGroupSeparator","t":8,"rt":$n[0].String,"fg":"numberGroupSeparator"},"s":{"a":2,"n":"set_NumberGroupSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"numberGroupSeparator"},"fn":"numberGroupSeparator"},{"a":2,"n":"NumberGroupSizes","t":16,"rt":$n[0].Array.type(System.Int32),"g":{"a":2,"n":"get_NumberGroupSizes","t":8,"rt":$n[0].Array.type(System.Int32),"fg":"numberGroupSizes"},"s":{"a":2,"n":"set_NumberGroupSizes","t":8,"p":[$n[0].Array.type(System.Int32)],"rt":$n[0].Void,"fs":"numberGroupSizes"},"fn":"numberGroupSizes"},{"a":2,"n":"PercentDecimalDigits","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_PercentDecimalDigits","t":8,"rt":$n[0].Int32,"fg":"percentDecimalDigits","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_PercentDecimalDigits","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"percentDecimalDigits"},"fn":"percentDecimalDigits"},{"a":2,"n":"PercentDecimalSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_PercentDecimalSeparator","t":8,"rt":$n[0].String,"fg":"percentDecimalSeparator"},"s":{"a":2,"n":"set_PercentDecimalSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"percentDecimalSeparator"},"fn":"percentDecimalSeparator"},{"a":2,"n":"PercentGroupSeparator","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_PercentGroupSeparator","t":8,"rt":$n[0].String,"fg":"percentGroupSeparator"},"s":{"a":2,"n":"set_PercentGroupSeparator","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"percentGroupSeparator"},"fn":"percentGroupSeparator"},{"a":2,"n":"PercentGroupSizes","t":16,"rt":$n[0].Array.type(System.Int32),"g":{"a":2,"n":"get_PercentGroupSizes","t":8,"rt":$n[0].Array.type(System.Int32),"fg":"percentGroupSizes"},"s":{"a":2,"n":"set_PercentGroupSizes","t":8,"p":[$n[0].Array.type(System.Int32)],"rt":$n[0].Void,"fs":"percentGroupSizes"},"fn":"percentGroupSizes"},{"a":2,"n":"PercentNegativePattern","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_PercentNegativePattern","t":8,"rt":$n[0].Int32,"fg":"percentNegativePattern","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_PercentNegativePattern","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"percentNegativePattern"},"fn":"percentNegativePattern"},{"a":2,"n":"PercentPositivePattern","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_PercentPositivePattern","t":8,"rt":$n[0].Int32,"fg":"percentPositivePattern","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_PercentPositivePattern","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"percentPositivePattern"},"fn":"percentPositivePattern"},{"a":2,"n":"PercentSymbol","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_PercentSymbol","t":8,"rt":$n[0].String,"fg":"percentSymbol"},"s":{"a":2,"n":"set_PercentSymbol","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"percentSymbol"},"fn":"percentSymbol"},{"a":2,"n":"PositiveInfinitySymbol","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_PositiveInfinitySymbol","t":8,"rt":$n[0].String,"fg":"positiveInfinitySymbol"},"s":{"a":2,"n":"set_PositiveInfinitySymbol","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"positiveInfinitySymbol"},"fn":"positiveInfinitySymbol"},{"a":2,"n":"PositiveSign","t":16,"rt":$n[0].String,"g":{"a":2,"n":"get_PositiveSign","t":8,"rt":$n[0].String,"fg":"positiveSign"},"s":{"a":2,"n":"set_PositiveSign","t":8,"p":[$n[0].String],"rt":$n[0].Void,"fs":"positiveSign"},"fn":"positiveSign"}]}; });
     $m("System.ComponentModel.DefaultValueAttribute", function () { return {"att":1048577,"a":2,"m":[{"a":2,"n":".ctor","t":1,"p":[$n[0].Boolean],"pi":[{"n":"value","pt":$n[0].Boolean,"ps":0}],"sn":"ctor"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Byte],"pi":[{"n":"value","pt":$n[0].Byte,"ps":0}],"sn":"$ctor1"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Char],"pi":[{"n":"value","pt":$n[0].Char,"ps":0}],"sn":"$ctor2"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Double],"pi":[{"n":"value","pt":$n[0].Double,"ps":0}],"sn":"$ctor3"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Int16],"pi":[{"n":"value","pt":$n[0].Int16,"ps":0}],"sn":"$ctor4"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Int32],"pi":[{"n":"value","pt":$n[0].Int32,"ps":0}],"sn":"$ctor5"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Int64],"pi":[{"n":"value","pt":$n[0].Int64,"ps":0}],"sn":"$ctor6"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Object],"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"sn":"$ctor7"},{"a":2,"n":".ctor","t":1,"p":[$n[0].SByte],"pi":[{"n":"value","pt":$n[0].SByte,"ps":0}],"sn":"$ctor8"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Single],"pi":[{"n":"value","pt":$n[0].Single,"ps":0}],"sn":"$ctor9"},{"a":2,"n":".ctor","t":1,"p":[$n[0].String],"pi":[{"n":"value","pt":$n[0].String,"ps":0}],"sn":"$ctor10"},{"a":2,"n":".ctor","t":1,"p":[$n[0].UInt16],"pi":[{"n":"value","pt":$n[0].UInt16,"ps":0}],"sn":"$ctor12"},{"a":2,"n":".ctor","t":1,"p":[$n[0].UInt32],"pi":[{"n":"value","pt":$n[0].UInt32,"ps":0}],"sn":"$ctor13"},{"a":2,"n":".ctor","t":1,"p":[$n[0].UInt64],"pi":[{"n":"value","pt":$n[0].UInt64,"ps":0}],"sn":"$ctor14"},{"a":2,"n":".ctor","t":1,"p":[Function,$n[0].String],"pi":[{"n":"type","pt":Function,"ps":0},{"n":"value","pt":$n[0].String,"ps":1}],"sn":"$ctor11"},{"ov":true,"a":2,"n":"Equals","t":8,"pi":[{"n":"obj","pt":$n[0].Object,"ps":0}],"sn":"equals","rt":$n[0].Boolean,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ov":true,"a":2,"n":"GetHashCode","t":8,"sn":"getHashCode","rt":$n[0].Int32,"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":3,"n":"SetValue","t":8,"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"sn":"setValue","rt":$n[0].Void,"p":[$n[0].Object]},{"v":true,"a":2,"n":"Value","t":16,"rt":$n[0].Object,"g":{"v":true,"a":2,"n":"get_Value","t":8,"rt":$n[0].Object,"fg":"Value"},"fn":"Value"},{"a":1,"n":"_value","t":4,"rt":$n[0].Object,"sn":"_value"}]}; });
     $m("System.ComponentModel.BrowsableAttribute", function () { return {"att":1048833,"a":2,"m":[{"a":2,"n":".ctor","t":1,"p":[$n[0].Boolean],"pi":[{"n":"browsable","pt":$n[0].Boolean,"ps":0}],"sn":"ctor"},{"ov":true,"a":2,"n":"Equals","t":8,"pi":[{"n":"obj","pt":$n[0].Object,"ps":0}],"sn":"equals","rt":$n[0].Boolean,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ov":true,"a":2,"n":"GetHashCode","t":8,"sn":"getHashCode","rt":$n[0].Int32,"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"Browsable","t":16,"rt":$n[0].Boolean,"g":{"a":2,"n":"get_Browsable","t":8,"rt":$n[0].Boolean,"fg":"Browsable","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"fn":"Browsable"},{"a":2,"n":"Default","is":true,"t":4,"rt":$n[14].BrowsableAttribute,"sn":"default","ro":true},{"a":2,"n":"No","is":true,"t":4,"rt":$n[14].BrowsableAttribute,"sn":"no","ro":true},{"a":2,"n":"Yes","is":true,"t":4,"rt":$n[14].BrowsableAttribute,"sn":"yes","ro":true},{"a":1,"n":"browsable","t":4,"rt":$n[0].Boolean,"sn":"browsable","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}}]}; });
+    $m("System.Collections.DictionaryEntry", function () { return {"att":1057033,"a":2,"at":[new System.SerializableAttribute()],"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Object,$n[0].Object],"pi":[{"n":"key","pt":$n[0].Object,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"sn":"$ctor1"},{"a":2,"n":"Key","t":16,"rt":$n[0].Object,"g":{"a":2,"n":"get_Key","t":8,"rt":$n[0].Object,"fg":"Key"},"s":{"a":2,"n":"set_Key","t":8,"p":[$n[0].Object],"rt":$n[0].Void,"fs":"Key"},"fn":"Key"},{"a":2,"n":"Value","t":16,"rt":$n[0].Object,"g":{"a":2,"n":"get_Value","t":8,"rt":$n[0].Object,"fg":"Value"},"s":{"a":2,"n":"set_Value","t":8,"p":[$n[0].Object],"rt":$n[0].Void,"fs":"Value"},"fn":"Value"},{"a":1,"n":"_key","t":4,"rt":$n[0].Object,"sn":"_key"},{"a":1,"n":"_value","t":4,"rt":$n[0].Object,"sn":"_value"}]}; });
     $m("System.Collections.IComparer", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"Compare","t":8,"pi":[{"n":"x","pt":$n[0].Object,"ps":0},{"n":"y","pt":$n[0].Object,"ps":1}],"sn":"System$Collections$IComparer$compare","rt":$n[0].Int32,"p":[$n[0].Object,$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Int32);}}]}; });
     $m("System.Collections.IEnumerator", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"MoveNext","t":8,"sn":"System$Collections$IEnumerator$moveNext","rt":$n[0].Boolean,"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ab":true,"a":2,"n":"Reset","t":8,"sn":"System$Collections$IEnumerator$reset","rt":$n[0].Void},{"ab":true,"a":2,"n":"Current","t":16,"rt":$n[0].Object,"g":{"ab":true,"a":2,"n":"get_Current","t":8,"rt":$n[0].Object,"fg":"System$Collections$IEnumerator$Current"},"fn":"System$Collections$IEnumerator$Current"}]}; });
     $m("System.Collections.IEqualityComparer", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"Equals","t":8,"pi":[{"n":"x","pt":$n[0].Object,"ps":0},{"n":"y","pt":$n[0].Object,"ps":1}],"sn":"System$Collections$IEqualityComparer$equals","rt":$n[0].Boolean,"p":[$n[0].Object,$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ab":true,"a":2,"n":"GetHashCode","t":8,"pi":[{"n":"obj","pt":$n[0].Object,"ps":0}],"sn":"System$Collections$IEqualityComparer$getHashCode","rt":$n[0].Int32,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Int32);}}]}; });
@@ -197,7 +198,6 @@
     $m("System.Collections.IDictionaryEnumerator", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"Entry","t":16,"rt":$n[7].DictionaryEntry,"g":{"ab":true,"a":2,"n":"get_Entry","t":8,"rt":$n[7].DictionaryEntry,"fg":"System$Collections$IDictionaryEnumerator$Entry"},"fn":"System$Collections$IDictionaryEnumerator$Entry"},{"ab":true,"a":2,"n":"Key","t":16,"rt":$n[0].Object,"g":{"ab":true,"a":2,"n":"get_Key","t":8,"rt":$n[0].Object,"fg":"System$Collections$IDictionaryEnumerator$Key"},"fn":"System$Collections$IDictionaryEnumerator$Key"},{"ab":true,"a":2,"n":"Value","t":16,"rt":$n[0].Object,"g":{"ab":true,"a":2,"n":"get_Value","t":8,"rt":$n[0].Object,"fg":"System$Collections$IDictionaryEnumerator$Value"},"fn":"System$Collections$IDictionaryEnumerator$Value"}]}; });
     $m("System.Collections.BitArray", function () { return {"nested":[$n[7].BitArray.BitArrayEnumeratorSimple],"att":1048833,"a":2,"m":[{"a":2,"n":".ctor","t":1,"p":[$n[0].Array.type(System.Boolean)],"pi":[{"n":"values","pt":$n[0].Array.type(System.Boolean),"ps":0}],"sn":"ctor"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Array.type(System.Byte)],"pi":[{"n":"bytes","pt":$n[0].Array.type(System.Byte),"ps":0}],"sn":"$ctor1"},{"a":2,"n":".ctor","t":1,"p":[$n[7].BitArray],"pi":[{"n":"bits","pt":$n[7].BitArray,"ps":0}],"sn":"$ctor2"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Int32],"pi":[{"n":"length","pt":$n[0].Int32,"ps":0}],"sn":"$ctor3"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Array.type(System.Int32)],"pi":[{"n":"values","pt":$n[0].Array.type(System.Int32),"ps":0}],"sn":"$ctor5"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Int32,$n[0].Boolean],"pi":[{"n":"length","pt":$n[0].Int32,"ps":0},{"n":"defaultValue","pt":$n[0].Boolean,"ps":1}],"sn":"$ctor4"},{"a":2,"n":"And","t":8,"pi":[{"n":"value","pt":$n[7].BitArray,"ps":0}],"sn":"And","rt":$n[7].BitArray,"p":[$n[7].BitArray]},{"a":2,"n":"Clone","t":8,"sn":"clone","rt":$n[0].Object},{"a":2,"n":"CopyTo","t":8,"pi":[{"n":"array","pt":Array,"ps":0},{"n":"index","pt":$n[0].Int32,"ps":1}],"sn":"copyTo","rt":$n[0].Void,"p":[Array,$n[0].Int32]},{"a":2,"n":"Get","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"sn":"Get","rt":$n[0].Boolean,"p":[$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"a":1,"n":"GetArrayLength","is":true,"t":8,"pi":[{"n":"n","pt":$n[0].Int32,"ps":0},{"n":"div","pt":$n[0].Int32,"ps":1}],"sn":"GetArrayLength","rt":$n[0].Int32,"p":[$n[0].Int32,$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"GetEnumerator","t":8,"sn":"GetEnumerator","rt":$n[7].IEnumerator},{"a":2,"n":"Not","t":8,"sn":"Not","rt":$n[7].BitArray},{"a":2,"n":"Or","t":8,"pi":[{"n":"value","pt":$n[7].BitArray,"ps":0}],"sn":"Or","rt":$n[7].BitArray,"p":[$n[7].BitArray]},{"a":2,"n":"Set","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0},{"n":"value","pt":$n[0].Boolean,"ps":1}],"sn":"Set","rt":$n[0].Void,"p":[$n[0].Int32,$n[0].Boolean]},{"a":2,"n":"SetAll","t":8,"pi":[{"n":"value","pt":$n[0].Boolean,"ps":0}],"sn":"SetAll","rt":$n[0].Void,"p":[$n[0].Boolean]},{"a":2,"n":"Xor","t":8,"pi":[{"n":"value","pt":$n[7].BitArray,"ps":0}],"sn":"Xor","rt":$n[7].BitArray,"p":[$n[7].BitArray]},{"a":2,"n":"Count","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_Count","t":8,"rt":$n[0].Int32,"fg":"Count","box":function ($v) { return Bridge.box($v, System.Int32);}},"fn":"Count"},{"a":2,"n":"IsReadOnly","t":16,"rt":$n[0].Boolean,"g":{"a":2,"n":"get_IsReadOnly","t":8,"rt":$n[0].Boolean,"fg":"IsReadOnly","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"fn":"IsReadOnly"},{"a":2,"n":"Item","t":16,"rt":$n[0].Boolean,"p":[$n[0].Int32],"i":true,"ipi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"g":{"a":2,"n":"get_Item","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"sn":"getItem","rt":$n[0].Boolean,"p":[$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"s":{"a":2,"n":"set_Item","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0},{"n":"value","pt":$n[0].Boolean,"ps":1}],"sn":"setItem","rt":$n[0].Void,"p":[$n[0].Int32,$n[0].Boolean]}},{"a":2,"n":"Length","t":16,"rt":$n[0].Int32,"g":{"a":2,"n":"get_Length","t":8,"rt":$n[0].Int32,"fg":"Length","box":function ($v) { return Bridge.box($v, System.Int32);}},"s":{"a":2,"n":"set_Length","t":8,"p":[$n[0].Int32],"rt":$n[0].Void,"fs":"Length"},"fn":"Length"},{"a":1,"n":"BitsPerByte","is":true,"t":4,"rt":$n[0].Int32,"sn":"BitsPerByte","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"BitsPerInt32","is":true,"t":4,"rt":$n[0].Int32,"sn":"BitsPerInt32","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"BytesPerInt32","is":true,"t":4,"rt":$n[0].Int32,"sn":"BytesPerInt32","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"_ShrinkThreshold","is":true,"t":4,"rt":$n[0].Int32,"sn":"_ShrinkThreshold","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"_version","t":4,"rt":$n[0].Int32,"sn":"_version","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"m_array","t":4,"rt":$n[0].Array.type(System.Int32),"sn":"m_array"},{"a":1,"n":"m_length","t":4,"rt":$n[0].Int32,"sn":"m_length","box":function ($v) { return Bridge.box($v, System.Int32);}}]}; });
     $m("System.Collections.BitArray.BitArrayEnumeratorSimple", function () { return {"td":$n[7].BitArray,"att":1048579,"a":1,"m":[{"a":4,"n":".ctor","t":1,"p":[$n[7].BitArray],"pi":[{"n":"bitarray","pt":$n[7].BitArray,"ps":0}],"sn":"ctor"},{"v":true,"a":2,"n":"MoveNext","t":8,"sn":"moveNext","rt":$n[0].Boolean,"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"a":2,"n":"Reset","t":8,"sn":"reset","rt":$n[0].Void},{"v":true,"a":2,"n":"Current","t":16,"rt":$n[0].Object,"g":{"v":true,"a":2,"n":"get_Current","t":8,"rt":$n[0].Object,"fg":"Current"},"fn":"Current"},{"a":1,"n":"bitarray","t":4,"rt":$n[7].BitArray,"sn":"bitarray"},{"a":1,"n":"currentElement","t":4,"rt":$n[0].Boolean,"sn":"currentElement","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"a":1,"n":"index","t":4,"rt":$n[0].Int32,"sn":"index","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":1,"n":"version","t":4,"rt":$n[0].Int32,"sn":"version","box":function ($v) { return Bridge.box($v, System.Int32);}}]}; });
-    $m("System.Collections.DictionaryEntry", function () { return {"att":1057033,"a":2,"at":[new System.SerializableAttribute()],"m":[{"a":2,"isSynthetic":true,"n":".ctor","t":1,"sn":"ctor"},{"a":2,"n":".ctor","t":1,"p":[$n[0].Object,$n[0].Object],"pi":[{"n":"key","pt":$n[0].Object,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"sn":"$ctor1"},{"a":2,"n":"Key","t":16,"rt":$n[0].Object,"g":{"a":2,"n":"get_Key","t":8,"rt":$n[0].Object,"fg":"Key"},"s":{"a":2,"n":"set_Key","t":8,"p":[$n[0].Object],"rt":$n[0].Void,"fs":"Key"},"fn":"Key"},{"a":2,"n":"Value","t":16,"rt":$n[0].Object,"g":{"a":2,"n":"get_Value","t":8,"rt":$n[0].Object,"fg":"Value"},"s":{"a":2,"n":"set_Value","t":8,"p":[$n[0].Object],"rt":$n[0].Void,"fs":"Value"},"fn":"Value"},{"a":1,"n":"_key","t":4,"rt":$n[0].Object,"sn":"_key"},{"a":1,"n":"_value","t":4,"rt":$n[0].Object,"sn":"_value"}]}; });
     $m("System.Collections.HashHelpers", function () { return {"att":1048960,"a":4,"s":true,"m":[{"a":2,"n":"ExpandPrime","is":true,"t":8,"pi":[{"n":"oldSize","pt":$n[0].Int32,"ps":0}],"sn":"ExpandPrime","rt":$n[0].Int32,"p":[$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"GetMinPrime","is":true,"t":8,"sn":"GetMinPrime","rt":$n[0].Int32,"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"GetPrime","is":true,"t":8,"pi":[{"n":"min","pt":$n[0].Int32,"ps":0}],"sn":"GetPrime","rt":$n[0].Int32,"p":[$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"IsPrime","is":true,"t":8,"pi":[{"n":"candidate","pt":$n[0].Int32,"ps":0}],"sn":"IsPrime","rt":$n[0].Boolean,"p":[$n[0].Int32],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"a":1,"n":"HashPrime","is":true,"t":4,"rt":$n[0].Int32,"sn":"HashPrime","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"MaxPrimeArrayLength","is":true,"t":4,"rt":$n[0].Int32,"sn":"MaxPrimeArrayLength","box":function ($v) { return Bridge.box($v, System.Int32);}},{"a":2,"n":"primes","is":true,"t":4,"rt":$n[0].Array.type(System.Int32),"sn":"primes","ro":true}]}; });
     $m("System.Collections.IDictionary", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"Add","t":8,"pi":[{"n":"key","pt":$n[0].Object,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"sn":"System$Collections$IDictionary$add","rt":$n[0].Void,"p":[$n[0].Object,$n[0].Object]},{"ab":true,"a":2,"n":"Clear","t":8,"sn":"System$Collections$IDictionary$clear","rt":$n[0].Void},{"ab":true,"a":2,"n":"Contains","t":8,"pi":[{"n":"key","pt":$n[0].Object,"ps":0}],"sn":"System$Collections$IDictionary$contains","rt":$n[0].Boolean,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ab":true,"a":2,"n":"GetEnumerator","t":8,"sn":"System$Collections$IDictionary$GetEnumerator","rt":$n[7].IDictionaryEnumerator},{"ab":true,"a":2,"n":"Remove","t":8,"pi":[{"n":"key","pt":$n[0].Object,"ps":0}],"sn":"System$Collections$IDictionary$remove","rt":$n[0].Boolean,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ab":true,"a":2,"n":"IsFixedSize","t":16,"rt":$n[0].Boolean,"g":{"ab":true,"a":2,"n":"get_IsFixedSize","t":8,"rt":$n[0].Boolean,"fg":"System$Collections$IDictionary$IsFixedSize","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"fn":"System$Collections$IDictionary$IsFixedSize"},{"ab":true,"a":2,"n":"IsReadOnly","t":16,"rt":$n[0].Boolean,"g":{"ab":true,"a":2,"n":"get_IsReadOnly","t":8,"rt":$n[0].Boolean,"fg":"System$Collections$IDictionary$IsReadOnly","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"fn":"System$Collections$IDictionary$IsReadOnly"},{"ab":true,"a":2,"n":"Item","t":16,"rt":$n[0].Object,"p":[$n[0].Object],"i":true,"ipi":[{"n":"key","pt":$n[0].Object,"ps":0}],"g":{"ab":true,"a":2,"n":"get_Item","t":8,"pi":[{"n":"key","pt":$n[0].Object,"ps":0}],"sn":"System$Collections$IDictionary$getItem","rt":$n[0].Object,"p":[$n[0].Object]},"s":{"ab":true,"a":2,"n":"set_Item","t":8,"pi":[{"n":"key","pt":$n[0].Object,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"sn":"System$Collections$IDictionary$setItem","rt":$n[0].Void,"p":[$n[0].Object,$n[0].Object]}},{"ab":true,"a":2,"n":"Keys","t":16,"rt":$n[7].ICollection,"g":{"ab":true,"a":2,"n":"get_Keys","t":8,"rt":$n[7].ICollection,"fg":"System$Collections$IDictionary$Keys"},"fn":"System$Collections$IDictionary$Keys"},{"ab":true,"a":2,"n":"Values","t":16,"rt":$n[7].ICollection,"g":{"ab":true,"a":2,"n":"get_Values","t":8,"rt":$n[7].ICollection,"fg":"System$Collections$IDictionary$Values"},"fn":"System$Collections$IDictionary$Values"}]}; });
     $m("System.Collections.IList", function () { return {"att":161,"a":2,"m":[{"ab":true,"a":2,"n":"Add","t":8,"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"tpc":0,"def":function (value) { return System.Array.add(this, value); },"rt":$n[0].Int32,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Int32);}},{"ab":true,"a":2,"n":"Clear","t":8,"tpc":0,"def":function () { return System.Array.clear(this); },"rt":$n[0].Void},{"ab":true,"a":2,"n":"Contains","t":8,"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"tpc":0,"def":function (value) { return System.Array.contains(this, value); },"rt":$n[0].Boolean,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},{"ab":true,"a":2,"n":"IndexOf","t":8,"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"tpc":0,"def":function (value) { return System.Array.indexOf(this, value, 0, null); },"rt":$n[0].Int32,"p":[$n[0].Object],"box":function ($v) { return Bridge.box($v, System.Int32);}},{"ab":true,"a":2,"n":"Insert","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"tpc":0,"def":function (index, value) { return System.Array.insert(this, index, value); },"rt":$n[0].Void,"p":[$n[0].Int32,$n[0].Object]},{"ab":true,"a":2,"n":"Remove","t":8,"pi":[{"n":"value","pt":$n[0].Object,"ps":0}],"tpc":0,"def":function (value) { return System.Array.remove(this, value); },"rt":$n[0].Void,"p":[$n[0].Object]},{"ab":true,"a":2,"n":"RemoveAt","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"sn":"System$Collections$IList$removeAt","rt":$n[0].Void,"p":[$n[0].Int32]},{"ab":true,"a":2,"n":"IsFixedSize","t":16,"rt":$n[0].Boolean,"g":{"ab":true,"a":2,"n":"get_IsFixedSize","t":8,"rt":$n[0].Boolean,"fg":"System$Collections$IList$IsFixedSize","box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}},"fn":"System$Collections$IList$IsFixedSize"},{"ab":true,"a":2,"n":"IsReadOnly","t":16,"rt":$n[0].Boolean,"g":{"ab":true,"a":2,"n":"get_IsReadOnly","t":8,"tpc":0,"def":function () { return System.Array.getIsReadOnly(this); },"rt":$n[0].Boolean,"box":function ($v) { return Bridge.box($v, System.Boolean, System.Boolean.toString);}}},{"ab":true,"a":2,"n":"Item","t":16,"rt":$n[0].Object,"p":[$n[0].Int32],"i":true,"ipi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"g":{"ab":true,"a":2,"n":"get_Item","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0}],"tpc":0,"def":function (index) { return System.Array.getItem(this, index); },"rt":$n[0].Object,"p":[$n[0].Int32]},"s":{"ab":true,"a":2,"n":"set_Item","t":8,"pi":[{"n":"index","pt":$n[0].Int32,"ps":0},{"n":"value","pt":$n[0].Object,"ps":1}],"tpc":0,"def":function (index, value) { return System.Array.setItem(this, index); },"rt":$n[0].Void,"p":[$n[0].Int32,$n[0].Object]}}]}; });
